[Rashmi]Added: RabbitMQ installed and started working on that and add…

+++ src/services/note.service.js
@@ -1,4 +1,4 @@
-import { result } from '@hapi/joi/lib/base';
+import { error, result } from '@hapi/joi/lib/base';
 import Note from '../models/note.model'; 
 import { client } from '../config/redis.js';
 
@@ -16,38 +16,47 @@ export const addNewNote = async (body) => {
 //get all notes
 export const getAllNotes = async (body) => {
     const data = await Note.find({UserId: body.UserId});
-    if(data){
+
+    if(data.length == null){
+      throw new error("There is no note with this user");
+    } else {
       await client.set('allNotes', JSON.stringify(data));
       return data;  
-    }
-    
+    }  
   };
+
+//get single note
+export const getNote = async (_id, UserId) => {
+  const data = await Note.findById({_id, UserId});
+  if(data == null){
+    throw new error("There is no note with this ID")
+  } else{
+    return data;
+  }
+};
   
 //update single note
 export const updateNote = async (_id, body) => {
-  const data = await Note.findByIdAndUpdate(
-    { _id: _id, UserId: body.UserId }, 
+  const forCheck = await Note.findById(_id);
+  if(forCheck == null){
+    throw new error("There is no note with this ID")
+  } else{
+    const data = await Note.findByIdAndUpdate({ _id: _id, UserId: body.UserId }, 
       body,
-    {
-      new: true
+      {
+        new: true
+      }
+    );
+    if(data){
+      await client.del('allnotes');
+      return data;
     }
-  );
-  if(data){
-    await client.del('allnotes');
-    return data;
-  }
-  
+  }  
 };
   
 //delete single note
 export const deleteNote = async (id, body) => {
   await Note.findByIdAndDelete({_id: id, UserId: body.UserId});
-};
-  
-//get single note
-export const getNote = async (_id, UserId) => {
-  const data = await Note.findById({_id, UserId});
-  return data;
 };
 
 //archive single note
